requirements.txt
langchain 
langchain_community 
langgraph 
langserve[all] 
langchain_google_genai
python-dotenv
langsmith